Turn pass into analysis.
A wrapper pass can be added later, if needed.
